+++ tests/compile_fail/error/forward_no_source_variant.rs
@@ -1,0 +1,19 @@
+use derive_more::Error;
+
+#[derive(Debug, Error)]
+enum Foo {
+    #[error(forward)]
+    Bar,
+    #[error(forward)]
+    Baz {
+        source: Box<dyn Error + Send + 'static>,
+    },
+}
+
+impl ::core::fmt::Display for Foo {
+    fn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {
+        write!(f, "")
+    }
+}
+
+fn main() {}